pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sliprun"
version = "0.1.0"
description = "BTC transaction builder with Ordinal inscriptions via Marathon Slipstream API"
authors = [{name = "ManjulShah", email = "manjul@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["bitcoin", "ordinals", "inscriptions", "slipstream", "mara"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31",
    "bitcoin-utils>=0.7.3",
    "click>=8.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.11",
    "responses>=0.23",
]

[project.scripts]
sliprun = "sliprun.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sliprun*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
